+++ packages/plugin-sound-cloud/src/toolbar/soundCloudURLInputModal.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { SoundCloudIcon } from '../icons';
 import classNames from 'classnames';
-import { mergeStyles, isSoundCloudUrl, SettingsPanelFooter, TextInput, CloseIcon } from 'wix-rich-content-common';
+import { mergeStyles, isSoundCloudUrl, SettingsPanelFooter, TextInput, CloseIcon, WixUtils } from 'wix-rich-content-common';
 import styles from '../../statics/styles/sound-cloud-url-input-modal.scss';
 
 export default class SoundCloudURLInputModal extends Component {
@@ -63,11 +63,11 @@ export default class SoundCloudURLInputModal extends Component {
 
     return (
       <div className={styles.container} data-hook="soundCloudUploadModal">
-        <CloseIcon className={classNames(styles.closeIcon)} onClick={() => this.onCloseRequested()} />
+          { !WixUtils.isMobile() && <CloseIcon className={classNames(styles.closeIcon)} onClick={() => this.onCloseRequested()} /> }
         <div role="heading" aria-labelledby="sound_cloud_modal_hdr" className={classNames(styles.header)}>
           <SoundCloudIcon className={classNames(styles.header_icon)} />
           <h3 id="sound_cloud_modal_hdr" className={styles.header_text}>
-            {t('SoundCloudUploadModal_Header')}
+              {(!WixUtils.isMobile()) ? t('SoundCloudUploadModal_Header') : t('SoundCloudUploadModal_Header_Mobile') }
           </h3>
         </div>
         <div className={styles.soundCloudUrlInputModal_textInput}>